Some work on index updates for GenericObjects

+++ src/org/zoodb/jdo/internal/Session.java
@@ -157,7 +157,9 @@ public class Session implements IteratorRegistry {
 		if (!cache.getDirtyGenericObjects().isEmpty()) {
     		for (GenericObject go: cache.getDirtyGenericObjects()) {
     			if (go.isDeleted() && !go.isNew()) {
-    				go.getClassDef().getProvidedContext().getDataDeleteSink().deleteGeneric(go);
+    				if (!go.checkPcDeleted()) {
+    					go.getClassDef().getProvidedContext().getDataDeleteSink().deleteGeneric(go);
+    				}
     			}
     		}
 		}
